Reorganize Evaluators Move the code specific to LogisticSimpleEvaluator to a new file logistic_simple_evaluator.jl, and code specific to LogisticCutoffEvaluator to new file logistic_cutoff_evaluator.jl. Include them in MSEP.jl, which seemed easier to understand than including them from evaluator.jl. Though maybe that would make more sense? Also not happy about mismatch file name and type name.
Moved common logic up to the abstract class, even though the default functions reference instance variables
that aren't in the abstract classes!  Also not sure about that, or the type declarations.

Futzed with the abstract hierarchy.

Defined zhat and zsimp, part of the new interface, for LogisticSimpleEvaluator, though
not for the cutoff, which may be broken as a result.  But it wasn't working anyway.

Made the worker processes use the new interface and be specific to LogisticSimpleEvaluator.

This is mostly intended to redistribute existing code, though it goes beyond that.