Output results to CSV. Add CSV package and update others. Weird error.
1. I got a conversion error current line 506,
        msi.estimated_iterations = round((sd/si.maxSD)^2)
can't convert to Int.  I think the error meant is couldn't convert NA to int, which I can
only imagine happening if there were <2 entries for the msep vector used to produce the
sd variable.  Yet this was around iteration 20, which may well have been the final iteration.
I was exec'ing the file in the VSCode Julia REPL.  I put some code in to check if
the sd was weird, mostly so I could add a breakpoint in it. But I never hit it again,
whether running in the REPL or the debugger.

2. Added CSV package to MSEP environment, though so far I have not used it.

3. Updated other packages, mostly to get the latest NamedArray 0.9.7; it was 0.9.6.
I don't think the change was material for me.

4.toCSV() writes the results to a CSV file.
Despite the name, it does not use the CSV package.
If I do use that package, CSV.write() is the key function.  That function does not know
how to handled NameArray, though that's somewhat moot since I have
to process it anyway to extract the values I'm interested in.

Successfully ran the function; waiting for feedback on the output format.

Note the use of the delightfully named enamerate.

5. Tightened the type of the msep vector inside MSEPInfo.
I noticed it was Any in the debugger; now Float64.