Fix estimation problems.  Remove absolute tolerance test in quadrature. Also fixed convergence check to use absolute values. But there may be problems from doing without the absolute tolerance; I added it originally to fix some issue.  But it might be the issue arose because I didn't use absolute values in my convergence test!
Results to quad3b.jld.

The show(r) added to quad_test.jl didn't work out; it just barfed out the numbers with no formatting:

PS C:\Users\rdboylan\Documents\BP\MSEP> date; julia --project -t16  -- src\quad_test.jl

Saturday, April 29, 2023 11:03:59 AM
[-2.4460519964103558 -2.446051996410113 -2.4460519964142162 -2.4460519964095644 -2.4460519964092815 -2.4460519964092726 -2.4460519964101226 -2.4460519964080687 -2.4460519964091607 -2.4460519964088356 -2.4460519964092344 -2.4460519964092913 -2.446051996409277; -1.3482345212221671 -1.3482345212221694 -1.348234521222168 -1.3482345212221565 -1.3482345212227305 -1.3482345212220757 -1.3482345212221862 -1.348234521222167 -1.3482345212221587 -1.3482345212210687 -1.348234521222342 -1.348234521222218 -1.348234521222132; -0.4767385578495229 -0.4767385578495203 -0.4767385578491825 -0.476738557849523 -0.4767385578495248 -0.4767385578495548 -0.4767385578494972 -0.47673855784952074 -0.4767385578495206 -0.4767385578495183 -0.4767385578495236 -0.4767385578495334 ......  # much more

Distributed results with this email:

I’ve tamed the misbehavior, but I’m not sure my solution is safe either.  First the results, then the caveats:
julia> rb = deserialize("quad3b.jld")
21×13 Named Matrix{Float64}
Y ╲ order │         3          4          5          6  …         12         13         14         15
──────────┼──────────────────────────────────────────────────────────────────────────────────────────
0         │  -2.44605   -2.44605   -2.44605   -2.44605  …   -2.44605   -2.44605   -2.44605   -2.44605
1         │  -1.34823   -1.34823   -1.34823   -1.34823      -1.34823   -1.34823   -1.34823   -1.34823
2         │ -0.476739  -0.476739  -0.476739  -0.476739     -0.476739  -0.476739  -0.476739  -0.476739
3         │  0.236797   0.236797   0.236797   0.236797      0.236797   0.236797   0.236797   0.236797
4         │  0.840592   0.840592   0.840592   0.840592      0.840592   0.840592   0.840592   0.840592
5         │   1.36715    1.36715    1.36715    1.36715       1.36715    1.36715    1.36715    1.36715
6         │   1.83851    1.83851    1.83851    1.83851       1.83851    1.83851    1.83851    1.83851
7         │   2.27003    2.27003    2.27003    2.27003       2.27003    2.27003    2.27003    2.27003
8         │   2.67283    2.67283    2.67283    2.67283       2.67283    2.67283    2.67283    2.67283
9         │   3.05533    3.05533    3.05533    3.05533       3.05533    3.05533    3.05533    3.05533
10        │   3.42422    3.42422    3.42422    3.42422       3.42422    3.42422    3.42422    3.42422
11        │   3.78515    3.78515    3.78515    3.78515       3.78515    3.78515    3.78515    3.78515
12        │   4.14317    4.14317    4.14317    4.14317       4.14317    4.14317    4.14317    4.14317
13        │   4.50312    4.50312    4.50312    4.50312       4.50312    4.50312    4.50312    4.50312
14        │      4.87       4.87       4.87       4.87          4.87       4.87       4.87       4.87
15        │   5.24928    5.24928    5.24928    5.24928       5.24928    5.24928    5.24928    5.24928
16        │   5.64741    5.64741    5.64741    5.64741       5.64741    5.64741    5.64741    5.64741
17        │    6.0724     6.0724     6.0724     6.0724        6.0724     6.0724     6.0724     6.0724
18        │   6.53473    6.53473    6.53473    6.53473       6.53473    6.53473    6.53473    6.53473
19        │   7.04874    7.04874    7.04874    7.04874       7.04874    7.04874    7.04874    7.04874
20        │   7.63489    7.63489    7.63489    7.63489  …    7.63489    7.63489    7.63489    7.63489

This time there are no entries > the values show above for 20, 7.63.

julia> maximum(rb)
7.634894714024786

julia> rb[13:21, 1:6]
9×6 Named Matrix{Float64}
Y ╲ order │       3        4        5        6        7        8
──────────┼─────────────────────────────────────────────────────
12        │ 4.14317  4.14317  4.14317  4.14317  4.14317  4.14317
13        │ 4.50312  4.50312  4.50312  4.50312  4.50312  4.50312
14        │    4.87     4.87     4.87     4.87     4.87     4.87
15        │ 5.24928  5.24928  5.24928  5.24928  5.24928  5.24928
16        │ 5.64741  5.64741  5.64741  5.64741  5.64741  5.64741
17        │  6.0724   6.0724   6.0724   6.0724   6.0724   6.0724
18        │ 6.53473  6.53473  6.53473  6.53473  6.53473  6.53473
19        │ 7.04874  7.04874  7.04874  7.04874  7.04874  7.04874
20        │ 7.63489  7.63489  7.63489  7.63489  7.63489  7.63489

The second table is the subset focused on in previous messages.

The change was the removal of an absolute tolerance of ~ 1.5e-8 [sqrt(machine epsilon)] as a stopping criterion for the adaptive quadrature.  This leaves only relative tolerance, which is much tighter when the thing being integrated is really small.
The accompanying graph shows a plot of the quantities being integrated for w (blue) and wz (lime) when Y=17.  logit(.85), rescaled for our mu and sigma, corresponds to ~7.5; as the graph shows, mixing with the normal density has pulled it in to around 6.
The maximum for w is around 2.5e-6, while for wz it is around 1.6e-5.  Since the curves are relatively wide, i.e., an interval of length 1 surrounding the mode has values close to the max at the ends, it seems even the smallest integral has a couple of orders magnitude on the absolute tolerance.  But apparently that wasn’t enough.  BTW the graph is completely scalable; I had to blow it up a lot to read the labels for the vertical scale.

The catch is that I’m pretty sure I inserted the absolute tolerance to avoid some other problem.  I don’t recall, and don’t seem to have noted, the details.
A clear problem is that if the quantity being integrated really is very close to 0 (probably this would be for wz) the procedure may decide it can’t satisfy the relative tolerance and signal an error.

Since the problems we’ve seen usually arise from misestimation of the denominator, integral of w, that really is close to 0 and for which inaccuracy has explosive consequences, it’s going to be challenging to meet all these conflicting demands.
Perhaps we should use an absolute tolerance, but only in the numerator.  But those are pretty small too…

Ross