quad3c output zhat for each possible sumY with clusters of 100.  Results to CEM
Add Tables to project so I can adapt NamedArray to CSV.
Add quad3c function to quad_test.jl to output desired analysis.  That file now using CSV and Tables.
Outputs to quad3c.{jld,csv}

I should have commented out the previous code for quad4(), but left it active.  So it ran again.

Email I sent with the csv:

@JN: cc'ing you b/c of a couple oddities I noticed, discussed in paragraphs 3 and 4.  In short, mu seems to imply 26.89 successes/100 trials, but for 27 successes zhat < 0.  And shrinkage seems to be pulling the top in more than the bottom.  Neither may be a surprise to you or @CM, but just in case it's interesting...

@CM Or maybe the attached is what you had in mind?  It gives the fit for every possible SumY from 0 through 100 for clusters of size 100.
The column headers in the first row give the order of the GK integrator.
The remaining rows are for SumY = 0..100, in that order (so SumY = Excel Row number - 2).
And the values in the body are zhat, as before  for zSQ(0.4).
As a quick check I took the standard deviation across orders for each row; all values were ~0 (I think 1e-12 was the biggest I saw).
mu = -1, sigma=1.  From quad_test.jl:quad3c().

Considering the outlandish probabilities involved for the extremes, I'm a bit surprised it seems to have handled them OK.
There is no true 0 value (according to these calculations and to my theory that asymmetry prevents it), with the smallest |zhat| for SumY=27 @ -0.006.
logistic(-1) = 0.2689, so an alternate theory is that since that does not give an expected integer # of success for 100 trials zhat can't be 0.
It's a little odd that zhat for 27, which is above the implied mode at 26.89, is not positive, but that may be from skew in the distn.

Also interesting that when one adds the intercept to zhat the min is -5.2 and the max is 4.8; I'm not sure why that isn't symmetric.  Maybe since the shrink is toward 0 (z scale), which is effectively -1 given mu (eta scale), there's more shrinkage at the top.